Refactor layer to separate module
